Fix Class Tangle
Between `LocalizedQueueConnectionFactory` and `NodeLocator` implementations.
