+++ spring-rabbit/src/main/java/org/springframework/amqp/rabbit/connection/FactoryFinder.java
@@ -1,0 +1,37 @@
+/*
+ * Copyright 2022 the original author or authors.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.springframework.amqp.rabbit.connection;
+
+/**
+ * Callback to determine the connection factory using the provided information.
+ *
+ * @author Gary Russell
+ * @since 2.4.8
+ */
+@FunctionalInterface
+public interface FactoryFinder {
+
+	/**
+	 * Locate or create a factory.
+	 * @param queueName the queue name.
+	 * @param node the node name.
+	 * @param nodeUri the node URI.
+	 * @return the factory.
+	 */
+	ConnectionFactory locate(String queueName, String node, String nodeUri);
+
+}
