+++ spring-rabbit/src/main/java/org/springframework/amqp/rabbit/connection/LocalizedQueueConnectionFactory.java
@@ -16,8 +16,6 @@
 
 package org.springframework.amqp.rabbit.connection;
 
-import java.net.URI;
-import java.net.URISyntaxException;
 import java.util.AbstractMap.SimpleImmutableEntry;
 import java.util.Arrays;
 import java.util.HashMap;
@@ -32,10 +30,8 @@ import org.springframework.amqp.AmqpException;
 import org.springframework.amqp.rabbit.support.RabbitExceptionTranslator;
 import org.springframework.beans.factory.DisposableBean;
 import org.springframework.core.io.Resource;
-import org.springframework.core.log.LogAccessor;
 import org.springframework.lang.Nullable;
 import org.springframework.util.Assert;
-import org.springframework.util.ClassUtils;
 
 /**
  * A {@link RoutingConnectionFactory} that determines the node on which a queue is located and
@@ -54,13 +50,6 @@ import org.springframework.util.ClassUtils;
  * @since 1.2
  */
 public class LocalizedQueueConnectionFactory implements ConnectionFactory, RoutingConnectionFactory, DisposableBean {
-
-	private static final boolean USING_WEBFLUX;
-
-	static {
-		USING_WEBFLUX = ClassUtils.isPresent("org.springframework.web.reactive.function.client.WebClient",
-				LocalizedQueueConnectionFactory.class.getClassLoader());
-	}
 
 	private final Log logger = LogFactory.getLog(getClass());
 
@@ -198,12 +187,7 @@ public class LocalizedQueueConnectionFactory implements ConnectionFactory, Routi
 		this.trustStore = trustStore;
 		this.keyStorePassPhrase = keyStorePassPhrase;
 		this.trustStorePassPhrase = trustStorePassPhrase;
-		if (USING_WEBFLUX) {
-			this.nodeLocator = new WebFluxNodeLocator();
-		}
-		else {
-			this.nodeLocator = new RestTemplateNodeLocator();
-		}
+		this.nodeLocator = ConnectionFactoryUtils.nodeLocator();
 	}
 
 	private static Map<String, String> nodesAddressesToMap(String[] nodes, String[] addresses) {
@@ -356,117 +340,6 @@ public class LocalizedQueueConnectionFactory implements ConnectionFactory, Routi
 	@Override
 	public void destroy() {
 		resetConnection();
-	}
-
-	/**
-	 * Used to obtain a connection factory for the queue leader.
-	 *
-	 * @param <T> the client type.
-	 * @since 2.4.8
-	 */
-	public interface NodeLocator<T> {
-
-		LogAccessor LOGGER = new LogAccessor(LogFactory.getLog(NodeLocator.class));
-
-		/**
-		 * Return a connection factory for the leader node for the queue.
-		 * @param adminUris an array of admin URIs.
-		 * @param nodeToAddress a map of node names to node addresses (AMQP).
-		 * @param vhost the vhost.
-		 * @param username the user name.
-		 * @param password the password.
-		 * @param queue the queue name.
-		 * @param factoryFunction an internal function to find or create the factory.
-		 * @return a connection factory, if the leader node was found; null otherwise.
-		 */
-		@Nullable
-		default ConnectionFactory locate(String[] adminUris, Map<String, String> nodeToAddress, String vhost,
-				String username, String password, String queue, FactoryFinder factoryFunction) {
-
-			T client = createClient(username, password);
-
-			for (int i = 0; i < adminUris.length; i++) {
-				String adminUri = adminUris[i];
-				if (!adminUri.endsWith("/api/")) {
-					adminUri += "/api/";
-				}
-				try {
-					String uri = new URI(adminUri)
-							.resolve("/api/queues/").toString();
-					Map<String, Object> queueInfo = restCall(client, uri, vhost, queue);
-					if (queueInfo != null) {
-						String node = (String) queueInfo.get("node");
-						if (node != null) {
-							String nodeUri = nodeToAddress.get(node);
-							if (nodeUri != null) {
-								close(client);
-								return factoryFunction.locate(queue, node, nodeUri);
-							}
-							if (LOGGER.isDebugEnabled()) {
-								LOGGER.debug("No match for node: " + node);
-							}
-						}
-					}
-					else {
-						throw new AmqpException("Admin returned null QueueInfo");
-					}
-				}
-				catch (Exception e) {
-					LOGGER.warn("Failed to determine queue location for: " + queue + " at: " +
-							adminUri + ": " + e.getMessage());
-				}
-			}
-			LOGGER.warn("Failed to determine queue location for: " + queue + ", using default connection factory");
-			close(client);
-			return null;
-		}
-
-		/**
-		 * Create a client for subsequent use.
-		 * @param userName the user name.
-		 * @param password the password.
-		 * @return the client.
-		 */
-		T createClient(String userName, String password);
-
-		/**
-		 * Close the client.
-		 * @param client the client.
-		 */
-		default void close(T client) {
-		}
-
-		/**
-		 * Retrieve a map of queue properties using the RabbitMQ Management REST API.
-		 * @param client the client.
-		 * @param baseUri the base uri.
-		 * @param vhost the virtual host.
-		 * @param queue the queue name.
-		 * @return the map of queue properties.
-		 * @throws URISyntaxException if the syntax is bad.
-		 */
-		@Nullable
-		Map<String, Object> restCall(T client, String baseUri, String vhost, String queue)
-				throws URISyntaxException;
-
-	}
-
-	/**
-	 * Callback to determine the connection factory using the provided information.
-	 * @since 2.4.8
-	 */
-	@FunctionalInterface
-	public interface FactoryFinder {
-
-		/**
-		 * Locate or create a factory.
-		 * @param queueName the queue name.
-		 * @param node the node name.
-		 * @param nodeUri the node URI.
-		 * @return the factory.
-		 */
-		ConnectionFactory locate(String queueName, String node, String nodeUri);
-
 	}
 
 }
